removed MERGED SassyToken with SassyJuggler
