+++ hardhat-SD/ignition/deployments/chain-534351/artifacts/SassyToken#SassyToken.dbg.json
@@ -1,0 +1,4 @@
+{
+  "_format": "hh-sol-dbg-1",
+  "buildInfo": "../build-info/c779ca26004f31f13a1614eb62d90bb2.json"
+}
