+++ hardhat-SD/ignition/deployments/chain-534351/deployed_addresses.json
@@ -1,0 +1,4 @@
+{
+  "SassyJuggler#SassyJuggler": "0xb05b5AE07921A856ED06D6Ad5e23434bE912DC47",
+  "SassyToken#SassyToken": "0x4B84A50d3B8944F1D091C7CaE22c2f728e379446"
+}
